+++ src/containers/ClassifContainer.py
@@ -1,5 +1,5 @@
 
-from src.Containers.SeqRecord import SeqRecord
+from src.containers.SeqRecord import SeqRecord
 
 
 class ClassifContainer:
