+++ src/reader_system/Fast5Reader.py
@@ -1,9 +1,9 @@
 
 from h5py import File
 
-from src.Containers.Fast5 import Fast5
+from src.containers.Fast5 import Fast5
 from src.simplify_read_id import simplify_read_id
-from src.ReaderSystem.FileReader import FileReader
+from src.reader_system.FileReader import FileReader
 
 
 class Fast5Reader(FileReader):
